+++ lib/i18n/app_localization_delegate.dart
@@ -14,7 +14,6 @@ class AppLocalizationsDelegate extends LocalizationsDelegate<Languages> {
   Future<Languages> load(Locale locale) => _load(locale);
 
   static Future<Languages> _load(Locale locale) async {
-    print(locale);
     switch (locale.languageCode) {
       case 'en':
         return LanguageEn();
